Fix plotter cutting a square every time
This fixes pmonta#17 and probably issue 20 in the that repo as
well. As far as I understand it, Ghostscript changed the way it inter-
prets PDF files, which breaks pstoedit. The project website of
[pstoedit](http://www.calvina.de/pstoedit) says "[eventually] a PDF file
will have to be converted to PostScript before it can be processed
by pstoedit", so this is what this commit does.

This commit fixes the command line tool as well as the GUI, the latter
getting a new input box to set the path to the Ghostscript executable.